Performance fixes
This is the first shot at improving the memory usage of searches
and helping the garbage collector do its job.  Additional
profiling should be done in the future, after the way searching
is done has been improved, to identify further areas to improve
the search performance and memory usage.

====

Instead of caching the jQuery requests used to retrieve the
followers after they have gone through, a boolean `true` or
`false` is now being cached.  The `getFollowers` method will
still return this cached boolean if the request for followers
has already been made, or the jQuery request promise if the
request has not been made before.

This also has the additional side-effect of limiting the number
of duplicated branches that are followed.  Now, instead of
following the branch along cached pathes, the branch is not
followed again.  Because of how the search is done, this should
not have an impact on calculating the chain, as branches other
than the first that are found will either have the same degree
as previous ones, or a larger degree.  This also greatly lowers
the size of the queue when searching along higher number of
degrees.

Because of the social nature of GitHub, and other social
networking-based websites, it is very likely that users will
share similar followers in higher degrees.  This reduces the
number of branches that must be checked, and as a result has a
significant speed increase when searching along higher degrees.

With this change, the user data can no longer be cached across
searches, which had previously resulted in large increases in
memory usage without the ability to clear the cache without
reloading the page.

====

Users are now shifted off of the queue using `splice(0, 1)` instead
of `shift`.  In large arrays, such as the one that we encounter
in higher degrees, this has a considerable performance gain when
processing it.

====

This also allows users to stop the search by hitting the search
button a second time while the search is in progress.  This will
stop the queue as soon as possible and clear up any data variables
used during the search.  This required adding a `searchChain`
boolean which will be `true` when a search is in progress and
`false` when a search should not be done.

====

This also adds some quick documentation to a few of the main
functions in a docstring format.  This should make it more clear
for those interested in contributing to the project to understand
what the methods do.