Work on the bidirectional search
This makes a few changes to the search so it will be able to better
handle working from both directions.

====

A function for determining the intersection between two arrays was
taken from StackOverflow [1] and is used to determine if any of the
users following the ending user are being followed by any of the
users who the starting user is following.  In more clear terms, this
allows us to work from both sides and find any middle ground.

Thanks to Paul S. for contributing the answer to Stack Overflow.
No changes were made to this function and it should appear in the
same state as the one found at that answer.

**Note:** This function is not compatible with IE < 10, and a
polyfill will be added in the future that will bring support for
at least IE 9.

====

Add arrays to keep track of users who have been discovered when
getting the followers.  An array has also been added to keep track
of users who have been discovered when going in reverse.  The same
has been done to keep track of the chains so they can be used to
determine where the intersection was.  Later, these chains will
be displayed to the user to allow them to visually see how the
starting user was able to get to the ending user.

When the scan starts, both the followers and following arrays are
initialized with the first and last users respectively, skipping
the checks for N=0, or when the start and last user are the same.

====

When each degree is being scanned, we no longer need to pass in
the ending user.  Because we are now using the intersection of
both searches, the ending user is not useful anymore.

A bit of the logic for passing the `reversedSearch` flag has also
been improve to prevent as much code duplication as possible.

====

The logic for displaying the chain has been adjusted to accept an
array of intersecting users, which will be used to go back and find
the pair of chains which can be combined together to get from the
starting user to the ending user.

====

[1]: http://stackoverflow.com/a/16227294/359284