Remove the submodules
We are removing the submodules as git cannot handle the latest
data set.  There are too many files, and the `stat` calls are
slowing everything down to a crawl.

Users must regenerate the data sets individually for testing or
request a compressed version of the data set to be sent to them.